refactor(questionRenderer): extract to util
